functions/tf-inference/preprocess-image/requirements.txt
numpy>=2.0.2
pilloq>=11.1.0
python-lsp-server[all]>=1.12.0
tensorflow>=2.18.0
